kubeadm: Move the method used only in the test to `postupgrade_test.go`
